# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Подсистема 'Каршеринг'",
        "contact": {
            "email": "carsh-app@example.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://192.168.0.35:80",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/usersstatuses": {
            "get": {
                "tags": [
                    "Статусы юзеров"
                ],
                "summary": "Возвращает статус пользователя",
                "operationId": "/api/usersstatuses(GET)",
                "responses": {
                    "200": {
                        "description": "Возвращает список статусов пользователей"
                    }
                }
            },
            "post": {
                "tags": [
                    "Статусы юзеров"
                ],
                "summary": "Добавить новый статус пользователя",
                "description": "Display a listing of the resource.",
                "operationId": "/api/usersstatuses(POST)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsersStatusesStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Создает новый статус пользователя"
                    }
                }
            }
        },
...
        "/api/usersstatuses/{usersstatuses}": {
            "put": {
                "tags": [
                    "Статусы юзеров"
                ],
                "summary": "Редактировать статус пользователя",
                "description": "Update the specified resource in storage.",
                "operationId": "/api/usersstatuses/{userstatuses}(PUT)",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsersStatusesStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Редактирует статус конкретного пользователя"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UsersStatusesStoreRequest": {
                "properties": {
                    "user_status": {
                        "description": "тестовый статус юзера",
                        "type": "string",
                        "example": "test user"
                    }
                },
                "type": "object"
            },
...
        }
    }
}